move handling of removing and selecting profiles to CreateDirectMessage component
